refactor(domain): move CurrentOrg from @batuda/controllers
- @batuda/email needs to read currentOrg.id from its ParticipantMatcher service, which would create a @batuda/controllers ↔ @batuda/email cycle (@batuda/controllers already imports EmailBlocks from @batuda/email). Domain is the leaf both can depend on without cycle. @batuda/controllers re-exports CurrentOrg so existing call sites compile unchanged.
