+++ packages/domain/src/index.ts
@@ -1,1 +1,2 @@
+export * from './current-org'
 export * from './schema/index'
